user userId in frontend for QOL

+++ backend/routes/posts.js
@@ -60,10 +60,15 @@ router.put('/:id', checkAuth, multer({ storage }).single('image'), (req, res, ne
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath
+    imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: 'Update successful!' });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update successful!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized!' });
+    }
   });
 });
 
@@ -100,8 +105,12 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    res.status(200).json({ message: 'Post deleted successfully!' });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Post deleted successfully!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized!' });
+    }
   });
 });
